+++ test/custom_widgets/custom_countdown_timer_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mobile_test/src/customWidgets/custom_countdown_timer.dart';
+import 'package:mobile_test/src/custom_widgets/custom_countdown_timer.dart';
 
 import '../screens/pump_widget_mock_setup.dart';
